chore: clean tracked artifacts and make the frontend type-check cleanly
Remove the 920 committed contracts/target build files plus stray empty/dead
files (IMPLEMENTATION_SUMMARY.md, portal Playground stubs, final-check,
test-fixes, the duplicate lowercase collaboration dir, and the stale
CourseGrid). Drop the redundant pnpm lockfile in favour of the committed npm
workspace lock and fix root scripts that still invoked pnpm.

Resolve all 150 frontend TypeScript errors (missing exports/interfaces, renamed
lucide icons, DOM/Web Audio/WebXR typings, Recharts formatters, IndexedDB
generics) and turn off `ignoreBuildErrors` so `next build` is type-checked
again. Also fix an SSR "window is not defined" crash in AuthContext that broke
prerendering of every /admin page. CI's frontend typecheck is now blocking.

+++ frontend/src/lib/bci/mlModel.ts
@@ -119,7 +119,9 @@ export class MLModel {
     try {
       const input = tf.tensor2d([features]);
       const prediction = this.model.predict(input) as tf.Tensor;
-      const values = await prediction.data();
+      // `.data()` is typed as a union of typed arrays; a float model always
+      // returns Float32Array, so narrow it for the helpers below.
+      const values = (await prediction.data()) as Float32Array;
 
       prediction.dispose();
       input.dispose();
